+++ typescript/utils/src/addresses.ts
@@ -1,6 +1,12 @@
 import { fromBech32, normalizeBech32, toBech32 } from '@cosmjs/encoding';
 import { PublicKey } from '@solana/web3.js';
 import { Wallet, utils as ethersUtils } from 'ethers';
+import {
+  addAddressPadding,
+  encode,
+  num,
+  validateAndParseAddress,
+} from 'starknet';
 
 import { isNullish } from './typeof.js';
 import { Address, HexString, ProtocolType } from './types.js';
@@ -8,6 +14,7 @@ import { assert } from './validation.js';
 
 const EVM_ADDRESS_REGEX = /^0x[a-fA-F0-9]{40}$/;
 const SEALEVEL_ADDRESS_REGEX = /^[a-zA-Z0-9]{36,44}$/;
+const STARKNET_ADDRESS_REGEX = /^(0x)?[0-9a-fA-F]{64}$/;
 
 const HEX_BYTES32_REGEX = /^0x[a-fA-F0-9]{64}$/;
 
@@ -24,10 +31,12 @@ const COSMOS_FACTORY_TOKEN_REGEX = new RegExp(
 const EVM_TX_HASH_REGEX = /^0x([A-Fa-f0-9]{64})$/;
 const SEALEVEL_TX_HASH_REGEX = /^[a-zA-Z1-9]{88}$/;
 const COSMOS_TX_HASH_REGEX = /^(0x)?[A-Fa-f0-9]{64}$/;
+const STARKNET_TX_HASH_REGEX = /^(0x)?[0-9a-fA-F]{64}$/;
 
 const EVM_ZEROISH_ADDRESS_REGEX = /^(0x)?0*$/;
 const SEALEVEL_ZEROISH_ADDRESS_REGEX = /^1+$/;
 const COSMOS_ZEROISH_ADDRESS_REGEX = /^[a-z]{1,10}?1[0]+$/;
+const STARKNET_ZEROISH_ADDRESS_REGEX = /^(0x)?0*$/;
 
 export const ZERO_ADDRESS_HEX_32 =
   '0x0000000000000000000000000000000000000000000000000000000000000000';
@@ -52,6 +61,10 @@ export function isCosmosIbcDenomAddress(address: Address): boolean {
   return IBC_DENOM_REGEX.test(address);
 }
 
+export function isAddressStarknet(address: Address) {
+  return STARKNET_ADDRESS_REGEX.test(address);
+}
+
 export function getAddressProtocolType(address: Address) {
   if (!address) return undefined;
   if (isAddressEvm(address)) {
@@ -60,6 +73,8 @@ export function getAddressProtocolType(address: Address) {
     return ProtocolType.Cosmos;
   } else if (isAddressSealevel(address)) {
     return ProtocolType.Sealevel;
+  } else if (isAddressStarknet(address)) {
+    return ProtocolType.Starknet;
   } else {
     return undefined;
   }
@@ -116,6 +131,15 @@ export function isValidAddressCosmos(address: Address) {
   }
 }
 
+export function isValidAddressStarknet(address: Address) {
+  try {
+    const isValid = address && validateAndParseAddress(address);
+    return !!isValid;
+  } catch {
+    return false;
+  }
+}
+
 export function isValidAddress(address: Address, protocol?: ProtocolType) {
   return routeAddressUtil(
     {
@@ -123,6 +147,7 @@ export function isValidAddress(address: Address, protocol?: ProtocolType) {
       [ProtocolType.Sealevel]: isValidAddressSealevel,
       [ProtocolType.Cosmos]: isValidAddressCosmos,
       [ProtocolType.CosmosNative]: isValidAddressCosmos,
+      [ProtocolType.Starknet]: isValidAddressStarknet,
     },
     address,
     false,
@@ -157,6 +182,14 @@ export function normalizeAddressCosmos(address: Address) {
   }
 }
 
+export function normalizeAddressStarknet(address: Address) {
+  if (isZeroishAddress(address)) return address;
+  try {
+    return validateAndParseAddress(address);
+  } catch {
+    return address;
+  }
+}
 export function normalizeAddress(address: Address, protocol?: ProtocolType) {
   return routeAddressUtil(
     {
@@ -183,6 +216,10 @@ export function eqAddressCosmos(a1: Address, a2: Address) {
   return normalizeAddressCosmos(a1) === normalizeAddressCosmos(a2);
 }
 
+export function eqAddressStarknet(a1: Address, a2: Address) {
+  return normalizeAddressStarknet(a1) === normalizeAddressStarknet(a2);
+}
+
 export function eqAddress(a1: Address, a2: Address) {
   const p1 = getAddressProtocolType(a1);
   const p2 = getAddressProtocolType(a2);
@@ -193,6 +230,7 @@ export function eqAddress(a1: Address, a2: Address) {
       [ProtocolType.Sealevel]: (_a1) => eqAddressSol(_a1, a2),
       [ProtocolType.Cosmos]: (_a1) => eqAddressCosmos(_a1, a2),
       [ProtocolType.CosmosNative]: (_a1) => eqAddressCosmos(_a1, a2),
+      [ProtocolType.Starknet]: (_a1) => eqAddressStarknet(_a1, a2),
     },
     a1,
     false,
@@ -212,6 +250,10 @@ export function isValidTransactionHashCosmos(input: string) {
   return COSMOS_TX_HASH_REGEX.test(input);
 }
 
+export function isValidTransactionHashStarknet(input: string) {
+  return STARKNET_TX_HASH_REGEX.test(input);
+}
+
 export function isValidTransactionHash(input: string, protocol: ProtocolType) {
   if (protocol === ProtocolType.Ethereum) {
     return isValidTransactionHashEvm(input);
@@ -221,6 +263,8 @@ export function isValidTransactionHash(input: string, protocol: ProtocolType) {
     return isValidTransactionHashCosmos(input);
   } else if (protocol === ProtocolType.CosmosNative) {
     return isValidTransactionHashCosmos(input);
+  } else if (protocol === ProtocolType.Starknet) {
+    return isValidTransactionHashStarknet(input);
   } else {
     return false;
   }
@@ -230,7 +274,8 @@ export function isZeroishAddress(address: Address) {
   return (
     EVM_ZEROISH_ADDRESS_REGEX.test(address) ||
     SEALEVEL_ZEROISH_ADDRESS_REGEX.test(address) ||
-    COSMOS_ZEROISH_ADDRESS_REGEX.test(address)
+    COSMOS_ZEROISH_ADDRESS_REGEX.test(address) ||
+    STARKNET_ZEROISH_ADDRESS_REGEX.test(address)
   );
 }
 
@@ -275,6 +320,11 @@ export function addressToBytesCosmos(address: Address): Uint8Array {
   return fromBech32(address).data;
 }
 
+export function addressToBytesStarknet(address: Address): Uint8Array {
+  const normalizedAddress = validateAndParseAddress(address);
+  return num.hexToBytes(normalizedAddress);
+}
+
 export function addressToBytes(
   address: Address,
   protocol?: ProtocolType,
@@ -285,6 +335,7 @@ export function addressToBytes(
       [ProtocolType.Sealevel]: addressToBytesSol,
       [ProtocolType.Cosmos]: addressToBytesCosmos,
       [ProtocolType.CosmosNative]: addressToBytesCosmos,
+      [ProtocolType.Starknet]: addressToBytesStarknet,
     },
     address,
     new Uint8Array(),
@@ -353,6 +404,11 @@ export function bytesToAddressCosmos(
   return toBech32(prefix, bytes);
 }
 
+export function bytesToAddressStarknet(bytes: Uint8Array): Address {
+  const hexString = encode.buf2hex(bytes);
+  return addAddressPadding(hexString);
+}
+
 export function bytesToProtocolAddress(
   bytes: Uint8Array,
   toProtocol: ProtocolType,
@@ -370,6 +426,8 @@ export function bytesToProtocolAddress(
     return bytesToAddressCosmos(bytes, prefix!);
   } else if (toProtocol === ProtocolType.CosmosNative) {
     return bytesToAddressCosmos(bytes, prefix!);
+  } else if (toProtocol === ProtocolType.Starknet) {
+    return bytesToAddressStarknet(bytes);
   } else {
     throw new Error(`Unsupported protocol for address ${toProtocol}`);
   }